Reworks the RoundRobinProcessor to be a subclass of Processor to allow for easier use of other processors

+++ Operating Systems/ProcessorManagement/src/processormanagement/Job.java
@@ -5,9 +5,9 @@ package processormanagement;
  * @author Gunnar
  */
 public class Job {
-    long sleepTime;
+    int sleepTime;
     
-    Job(long sleepTime) {
+    Job(int sleepTime) {
         this.sleepTime = sleepTime;
     }
 }
